% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploted.R
\name{lineEd}
\alias{lineEd}
\title{Plot the education level of a given county.}
\usage{
lineEd(stateshort, countyname)
}
\arguments{
\item{stateshort}{2 character abbreviation of states in US.}

\item{countyname}{County name.}
}
\description{
Plot the education level of a given county.
}
\examples{
lineEd("IA", "Story County")
mapstate(edlevel="Bachelor's degree or higher")
mapcounty(edlevel="Bachelor's degree or higher",stateshort="IA")
}
