% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploted.R
\name{mapcounty}
\alias{mapcounty}
\title{Map the education level of a given state.}
\usage{
mapcounty(vtype = "percent", levelint, stateshort, limit = 28)
}
\arguments{
\item{vtype}{"count" or "percent".}

\item{levelint}{Education level: 1="Less than a high school diploma", 2="High school diploma only", 3="Some college (1-3 years)" or 4="Bachelor's degree or higher".}

\item{stateshort}{2 character abbreviation of states in US.}

\item{limit}{Label county names whose percentage >= limit, defalut limit = 28}
}
\description{
Map the education level of a given state.
}
\examples{
lineEd("IA", "Story County")
mapstate(edlevel="Bachelor's degree or higher")
mapcounty(edlevel="Bachelor's degree or higher",stateshort="IA")
}
