% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploted.R
\name{mapstate}
\alias{mapstate}
\title{Map the education level of all states.}
\usage{
mapstate(vtype = "percent", levelint)
}
\arguments{
\item{vtype}{"count" or "percent".}

\item{levelint}{Education level: 1="Less than a high school diploma", 2="High school diploma only", 3="Some college (1-3 years)" or 4="Bachelor's degree or higher".}
}
\description{
Map the education level of all states.
}
\examples{
lineEd("IA", "Story County")
mapstate(edlevel="Bachelor's degree or higher")
mapcounty(edlevel="Bachelor's degree or higher",stateshort="IA")
}
