% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploted.R
\name{showcounty}
\alias{showcounty}
\title{Show list of counties of a given state.}
\usage{
showcounty(stateshort)
}
\arguments{
\item{stateshort}{2 character abbreviation of states in US. use "state.abb" to show abbreviation of all states}
}
\description{
Show list of counties of a given state.
}
\examples{
showcounty("IA")

}
